# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SIAKAD GERBANG",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reference/admission-pathways": {
            "get": {
                "tags": [
                    "Reference/AdmissionPathway"
                ],
                "summary": "List all AdmissionPathway",
                "operationId": "658bd7ec3bda1763e9615f57e1fea5d2",
                "responses": {
                    "200": {
                        "description": "List of AdmissionPathway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdmissionPathwayListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reference/AdmissionPathway"
                ],
                "summary": "Create a new AdmissionPathway",
                "operationId": "8abc4588394118402bc2e209154009ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "object",
                                        "example": "example name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "AdmissionPathway created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdmissionPathwayResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/reference/admission-pathways/{id}": {
            "get": {
                "tags": [
                    "Reference/AdmissionPathway"
                ],
                "summary": "Get AdmissionPathway detail by ID",
                "operationId": "d067f6007356b49b25fb198c8f89502d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AdmissionPathway ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdmissionPathway detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdmissionPathwayResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "AdmissionPathway not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reference/AdmissionPathway"
                ],
                "summary": "Update an existing AdmissionPathway",
                "operationId": "1abbbb396dc532457a17129d56f2a84b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AdmissionPathway ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "object",
                                        "example": "example name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "AdmissionPathway updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdmissionPathwayResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "AdmissionPathway not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reference/AdmissionPathway"
                ],
                "summary": "Delete a AdmissionPathway by ID",
                "operationId": "a174c34c76e6d270e1df2fc2ec79ad85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "AdmissionPathway ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdmissionPathway deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Data deleted",
                                    "data": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "AdmissionPathway not found"
                    }
                }
            }
        },
...
        "/api/reference/digit-semesters": {
            "get": {
                "tags": [
                    "Reference/DigitSemester"
                ],
                "summary": "List all DigitSemester",
                "operationId": "35bb4a69d36dcd3312bb3880d9a9d4b3",
                "responses": {
                    "200": {
                        "description": "List of DigitSemester",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DigitSemesterListResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reference/DigitSemester"
                ],
                "summary": "Create a new DigitSemester",
                "operationId": "8f8cb94748cfa6a864f86ce55f19017f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "DigitSemester created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DigitSemesterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/reference/digit-semesters/{id}": {
            "get": {
                "tags": [
                    "Reference/DigitSemester"
                ],
                "summary": "Get DigitSemester detail by ID",
                "operationId": "39d7176768adda18ed5837809f6cd76f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "DigitSemester ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DigitSemester detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DigitSemesterResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "DigitSemester not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reference/DigitSemester"
                ],
                "summary": "Update an existing DigitSemester",
                "operationId": "2daa63f0388ab44a19210f79e8fbf2bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "DigitSemester ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "DigitSemester updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DigitSemesterResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "DigitSemester not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reference/DigitSemester"
                ],
                "summary": "Delete a DigitSemester by ID",
                "operationId": "7225c9f18809fa0254e2a76a0609bdaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "DigitSemester ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DigitSemester deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": "success",
                                    "message": "Data deleted",
                                    "data": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "DigitSemester not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdmissionPathway": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "properties": {
                            "en": {
                                "type": "string",
                                "example": "Pending"
                            },
                            "id": {
                                "type": "string",
                                "example": "Menunggu"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "AdmissionPathwayResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Admission Pathway detail"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AdmissionPathway"
                    }
                },
                "type": "object"
            },
            "AdmissionPathwayListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "List of admission pathways"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdmissionPathway"
                        }
                    }
                },
                "type": "object"
            },
...
            "DigitSemester": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "digit": {
                        "type": "integer",
                        "example": 1
                    },
                    "note": {
                        "type": "json",
                        "example": "{}"
                    }
                },
                "type": "object"
            },
            "DigitSemesterResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Digit semester detail"
                    },
                    "data": {
                        "$ref": "#/components/schemas/DigitSemester"
                    }
                },
                "type": "object"
            },
            "DigitSemesterListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "List of digit semesters"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DigitSemester"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}